+++ tests/integration/Section.test.js
@@ -151,7 +151,7 @@ test('tables - get - number', (t) => {
 test('templates - get', (t) => {
   let str = fs.readFileSync(path.join(dir, '../', 'cache', 'United-Kingdom.txt'), 'utf-8')
   let sec = wtf(str).section(0)
-  const expected = [148, 195, 54, 49, 296, 176, 182, 399, 70, 97, 147, 78, 23, 18, 18, 21]
+  const expected = [148, 195, 54, 49, 176, 182, 399, 70, 97, 147, 78, 23, 18, 18, 21]
   t.deepEqual(
     sec.templates().map((s) => JSON.stringify(s.json()).length),
     expected,
@@ -178,7 +178,7 @@ test('templates - get - string', (t) => {
 test('infoboxes - get', (t) => {
   let str = fs.readFileSync(path.join(dir, '../', 'cache', 'United-Kingdom.txt'), 'utf-8')
   let sec = wtf(str).section(0)
-  const expected = [33]
+  const expected = [26, 33]
   t.deepEqual(
     sec.infoboxes().map((s) => JSON.stringify(s).length),
     expected,
@@ -190,7 +190,7 @@ test('infoboxes - get', (t) => {
 test('infoboxes - get - number', (t) => {
   let str = fs.readFileSync(path.join(dir, '../', 'cache', 'United-Kingdom.txt'), 'utf-8')
   let sec = wtf(str).section(0)
-  t.equal(JSON.stringify(sec.infobox()).length, 33, 'the infoboxes should equal the expected')
+  t.equal(JSON.stringify(sec.infobox()).length, 26, 'the infoboxes should equal the expected')
   t.end()
 })